moved Base/FlagDataFixture.php to Demo/FlagDataFixture.php because the shop can work without flags by default
